first pass of float-based simulation including: - use of single-precision ODE; - new algorithm for picking a new origin, ReviseOrigin, which takes into account the distance from nearest piece of formation geometry so resets don't occur continuously when flying high and fast; - shift away from engine-specific geometric types in favor or Abs/Rel variants; - deprecates the hysteresis on recalculating node scores for smoother node churn; - many bugs and glitches

+++ src/gfx/defs.h
@@ -36,9 +36,8 @@ namespace gfx
 		};
 	}
 	
-	// Geometric types currently use double and are converted 
-	// inside - or close to - the graphics API.
-	typedef double Scalar;
+	// geometric types
+	typedef float Scalar;
 	typedef geom::Vector<Scalar, 3> Vector3;
 	typedef geom::Vector<Scalar, 4> Vector4;
 	typedef geom::Vector<Scalar, 2> RenderRange;	// near/far z
